fix: honor large Prometheus rate-limit bursts
Widen the per-IP Prometheus token bucket to u64 milli-tokens so
--prom-rate-limit-burst values above u32::MAX / 1000 do not silently
saturate below the configured burst.

Add a focused regression for large u32 burst values.

+++ crates/varta-watch/src/exporter/mod.rs
@@ -352,8 +352,10 @@ const PROM_IP_STATE_SWEEP_INTERVAL: Duration = Duration::from_secs(10);
 #[derive(Clone, Copy, Debug)]
 struct PromIpState {
     /// Tokens available (fractional, scaled by 1000 to avoid floats).
-    /// Each accepted connection consumes 1000 milli-tokens.
-    tokens_milli: u32,
+    /// Each accepted connection consumes 1000 milli-tokens. This is wider
+    /// than the configured token count because `rate_burst` accepts `u32`
+    /// whole tokens and the internal scale factor must not lower that cap.
+    tokens_milli: u64,
     /// Wall-clock instant at which `tokens_milli` was last refilled.
     last_refill: Instant,
     /// Most recent connection from this IP — used for stale eviction.
